fix: Refactoring the code
