+++ src/components/explore/search-bar.tsx
